requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.6.4.0
celery==4.4.7
Deprecated==1.2.13
Django==3.2.12
django-crontab==0.7.1
django-filter==21.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
importlib-metadata==4.11.2
kombu==4.6.11
packaging==21.3
pkg_resources==0.0.0
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2021.3
redis==4.1.0
sqlparse==0.4.2
typing_extensions==4.1.1
vine==1.3.0
wrapt==1.13.3
zipp==3.7.0
